server/requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
python-dotenv==1.0.1
requests==2.32.3
youtube-transcript-api==0.6.2
faiss-cpu==1.8.0.post1
numpy==1.26.4
scikit-learn==1.5.2
openai==1.51.2
PyYAML==6.0.2
yt-dlp==2024.8.6
